Json library is now working. Updated readme to show how to download libraries if necessary.
