--- Docker_Mqtt/files/cheese.txt
@@ -1,6 +1,0 @@
-Hello my name is Rodrigo and I am wondering if I can send a file
-across the network,
-
-Cheers, from the otherside
-
-Rod
