--- Docker_Mqtt/files/encoded_file.txt
@@ -1,3 +1,0 @@
-SGVsbG8gbXkgbmFtZSBpcyBSb2RyaWdvIGFuZCBJIGFtIHdvbmRlcmluZyBpZiBJIGNhbiBzZW5k
-IGEgZmlsZQphY3Jvc3MgdGhlIG5ldHdvcmssCgpDaGVlcnMsIGZyb20gdGhlIG90aGVyc2lkZQoK
-Um9k
